+++ src/fonduer/features/feature_libs/structural_features.py
@@ -1,4 +1,4 @@
-from typing import Dict, Iterator, List, Set, Tuple
+from typing import Dict, Iterator, List, Set, Tuple, Union
 
 from fonduer.candidates.models import Candidate
 from fonduer.candidates.models.span_mention import SpanMention, TemporarySpanMention
@@ -23,7 +23,7 @@ binary_strlib_feats: Dict[str, Set[Tuple[str, int]]] = {}
 
 
 def extract_structural_features(
-    candidates: List[Candidate]
+    candidates: Union[Candidate, List[Candidate]],
 ) -> Iterator[Tuple[int, str, int]]:
     """Extract structural features.