+++ tests/features/test_features.py
@@ -12,16 +12,16 @@ from tests.shared.hardware_matchers import part_matcher, temp_matcher
 
 logger = logging.getLogger(__name__)
 DB = "feature_test"
+# Use 127.0.0.1 instead of localhost (#351)
+CONN_STRING = f"postgresql://127.0.0.1:5432/{DB}"
 
 
-def test_feature_extraction(caplog):
+def test_feature_extraction():
     """Test extracting candidates from mentions from documents."""
-    caplog.set_level(logging.INFO)
-
     PARALLEL = 1
 
     max_docs = 1
-    session = Meta.init("postgresql://localhost:5432/" + DB).Session()
+    session = Meta.init(CONN_STRING).Session()
 
     docs_path = "tests/data/html/"
     pdf_path = "tests/data/pdf/"